Extract api/routes/ — thin main.py to startup-only
Mechanical move of route handlers out of main.py into a new
backend/api/routes/ package. No behavior changes — URLs, status
codes, response shapes, and background-task wiring all identical.

- api/routes/health.py — GET /health, GET /debug/config
- api/routes/intake.py — POST /api/upload (_intake_error_to_http
  moves with it; it's only used by the upload handler)
- api/routes/auth.py — GET /authorize, GET /oauth2callback
- main.py — keeps app creation, CORS, startup hooks, exception
  handlers, and _json_error; registers routes via app.include_router

main.py goes from 204 lines to ~80. Route files use APIRouter();
no FastAPI app instance leaks into them.

Closes #155. Sub-item of the #74 backend refactor (PDF PR 5).
